Progress 80%

+++ database/migrations/2023_09_25_180311_add_columns_to_users_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('id_card')->nullable()->after('email');
             $table->boolean('custom_fee')->default(false)->after('id_card');
             $table->integer('fee')->default(0)->after('custom_fee');
-            $table->integer('ballance')->default(0)->after('fee');
+            $table->decimal('ballance', 8, 2)->default(0)->after('fee');
             $table->string('referal')->nullable()->default('system')->after('ballance');
             $table->boolean('is_active')->default(false)->after('referal');
             $table->boolean('is_admin')->default(false)->after('is_active');
